scraper/pyproject.toml
[project]
name = "chapssal-scraper"
version = "0.1.0"
description = ""
authors = [
    {name = "rokrokss",email = "q0115643@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "google-genai (>=1.12.1,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "supabase (>=2.15.1,<3.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "psycopg-c (>=3.2.6,<4.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
